# Reduce device synchronisation