@RAMitchell 2020-05-06
@ShvetsKS the use of vector<bool> in DenseColumn seems to be causing test failures on my system and on CI.