@ShvetsKS 2020-05-06
@RAMitchell I don't have any objections to use `feature_offset_` instead of displaced iterator if we don't hide a real problem.
But it's not clear what is exact root cause of iterator crash.