@RAMitchell 2020-05-06
@ShvetsKS See https://xgboost-ci.net/blue/organizations/jenkins/xgboost/detail/PR-5633/2/pipeline/108

Boolean vector is a strange thing. My guess is that the iterator contains a reference to the original and something happens when the iterator gets destructed.