@ShvetsKS 2020-05-06
@RAMitchell  thanks for fixing this issue. It was just interesting for me why we didn't see this fail before this PR.