@trivialfis 2020-05-06
> Boolean vector

I don't mind forbidding its usage.