@RAMitchell 2020-05-06
Benchmarks for this PR:
| Platform | GPU    | Grow policy | Depth | Before PR (s) | This PR (s) | Speedup     |
|----------|--------|-------------|-------|---------------|-------------|-------------|
| Windows  | 1080ti | depthwise   | 6     | 9.7           | 8.5         | 1.141176471 |
| Windows  | 1080ti | depthwise   | 8     | 22.5          | 16.7        | 1.347305389 |
| Windows  | 1080ti | lossguide   | 6     | 9.7           | 9.2         | 1.054347826 |
| Windows  | 1080ti | lossguide   | 8     | 21.7          | 20.3        | 1.068965517 |
| Ubuntu   | V100   | depthwise   | 6     | 4.9           | 4.6         | 1.065217391 |
| Ubuntu   | V100   | depthwise   | 8     | 9.3           | 8.1         | 1.148148148 |
| Ubuntu   | V100   | lossguide   | 6     | 4.8           | 4.8         | 1           |
| Ubuntu   | V100   | lossguide   | 8     | 9.3           | 8.8         | 1.056818182 |