@RAMitchell 2020-05-06
@trivialfis they may still need it as the difference in memory consumption is 8x vs vector<int8_t>.

@hcho3 we should cherry pick the changes to src/common/column_matrix.h into the release. Let me move it to a separate PR.